+++ firebase-ai/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
-
+* [feature] Added support for configuring the "thinking" budget when using Gemini
+  2.5 series models. (#6990)
+* [feature] **Breaking Change**: Add support for Grounding with Google Search (#7042).
+    * **Action Required:** Update all references of `groundingAttributions`, `webSearchQueries`, `retrievalQueries` in `GroundingMetadata` to be non-optional.
 
 # 16.2.0
 * [changed] Deprecate the `totalBillableCharacters` field (only usable with pre-2.0 models). (#7042)
@@ -31,4 +34,3 @@
 
  Note: This feature is in Public Preview, which means that it is not subject to any SLA or
  deprecation policy and could change in backwards-incompatible ways.
-